refactor: extract all functions into files
