+++ src/mapProps.ts
@@ -1,0 +1,7 @@
+import type { ReactTestRendererTreeFixed } from './types';
+
+export function mapProps(props: ReactTestRendererTreeFixed['props']) {
+  const { children, ...propsWithoutChildren } = props;
+
+  return propsWithoutChildren;
+}
